requirements.txt
torch
transformers
accelerate
datasets
bitsandbytes
peft
python-chess
trl
huggingface_hub
